requirements.txt
anyio==3.6.2
autopep8==2.0.1
boto3==1.26.69
botocore==1.29.74
certifi==2022.12.7
charset-normalizer==3.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
parse==1.6.5
pycodestyle==2.10.0
pynamodb==5.4.0
pyparser==1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-i18n==0.3.9
python-telegram-bot==20.0
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
urllib3==1.26.14
marshmallow==3.21.3
marshmallow-dataclass==8.6.1